MistyRoboticsFullCode-EducatorAwardPresentation.js
This program was utilized for the presentation of the Educator of Distinction Award given by the Coca-Cola Scholars Foundation. Misty the Robot was programmed to drag a cart with the award, present the award to the awardee, move/act in human-like ways (beating pulse light like a heart, natural head movements and eye blinking, etc.) , and wave goodbye as she drove away. Currently, the facial recognition section is commented out as it was not used in the actual presentation and needs to be explored further before official implementation.